Separate out realphone and relaynumber API hooks
